% Auto-generated: do not edit by hand
\name{dashFullscreen}

\alias{dashFullscreen}

\title{DashFullscreen component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dashFullscreen(children=NULL, id=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number | unnamed list. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

